Refactor kernel constants and system halt logic
Moved kernel constants and Limine requests to a new consts.rs module. Relocated system halt logic to utils/system.rs and updated references throughout the kernel. Printing macros and framebuffer message display were moved to scrolling_text.rs for better modularity.
