+++ kernel/src/utils/mod.rs
@@ -1,2 +1,3 @@
 pub mod bsod;
 pub mod color_test;
+pub mod system;
