+++ kernel/src/utils/system.rs
@@ -1,0 +1,9 @@
+
+/// Safe system halt
+pub fn halt_system() -> ! {
+    loop {
+        unsafe {
+            core::arch::asm!("hlt");
+        }
+    }
+}
